v.1.11.0 Option to jump weekends when use the arrows to jump to next/prev day. UI now show the name of weekday and nº week of the wear in start/end dates.
Now, on Holded Check-In:
Now can specify in option the projectID of log on Toggl, created at check-in in holded.
If there are some log already running on toggl, don't stop it
Option to avoid already running toggl log, and force start the default task of check-in.

Variable to include a invisible "firm" in worklog description in Jira, a invisible char, for analytics proposes (default disabled)